+++ releasenotes.md
@@ -1,4 +1,5 @@
 Whats New in v10.7
 --------------------------
+- Choose which supplier you updating the stock when making a formula
 - Show available stock from all the available suppliers when making a formula, instead of the preferred one only
 - This release may include more changes, for full details please refer to the CHANGELOG